requirements.txt
python-dotenv==1.0.0
ollama==0.1.7
langchain==0.1.10
langchain-google-genai==0.0.11
google-generativeai==0.3.2